remove old code and rewrite paths
